% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{GenData_FiveCov_outlier}
\alias{GenData_FiveCov_outlier}
\title{Generate Simulated Data with Outliers in X1}
\usage{
GenData_FiveCov_outlier(nprim = 300, naux = 1500, setting = 1, Dist = "Norm")
}
\arguments{
\item{nprim}{Integer. Sample size for primary data. Default is 300.}

\item{naux}{Integer. Sample size for auxiliary data. Default is 1500.}

\item{setting}{Integer from 1 to 4. Specifies simulation setting.}

\item{Dist}{Character. Distribution of the third covariate: "Norm" (normal) or "Unif" (uniform).}
}
\value{
A list with primary and auxiliary datasets including outliers.
}
\description{
Same as \code{GenData_FiveCov} but adds outliers in X1 for the primary dataset.
}
