% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{GetBeta}
\alias{GetBeta}
\title{Estimate Regression Coefficients via Weighted Quantile Regression}
\usage{
GetBeta(data)
}
\arguments{
\item{data}{Data frame containing covariates X1–X5 and time, delta.}
}
\value{
Estimated regression coefficients as a numeric vector.
}
\description{
Estimate Regression Coefficients via Weighted Quantile Regression
}
