% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{GetBIC}
\alias{GetBIC}
\title{Compute BIC for Model Selection}
\usage{
GetBIC(data, beta_a, eta, lambda, cutoff = 1e-05)
}
\arguments{
\item{data}{Data frame.}

\item{beta_a}{Auxiliary coefficient vector.}

\item{eta}{Adjustment vector.}

\item{lambda}{Regularization parameter.}

\item{cutoff}{Threshold for non-zero coefficients. Default is 1e-5.}
}
\value{
BIC value.
}
\description{
Compute BIC for Model Selection
}
