% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{GetEta}
\alias{GetEta}
\title{Estimate Adjustment Term Eta}
\usage{
GetEta(primdata, beta_a)
}
\arguments{
\item{primdata}{Primary data frame.}

\item{beta_a}{Auxiliary model coefficients.}
}
\value{
Estimated adjustment vector eta.
}
\description{
Estimate Adjustment Term Eta
}
