% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{GetEtalad}
\alias{GetEtalad}
\title{Estimate Eta Using LAD Regression}
\usage{
GetEtalad(primdata, beta_a)
}
\arguments{
\item{primdata}{Primary data frame.}

\item{beta_a}{Auxiliary model coefficients.}
}
\value{
Estimated eta vector from LAD regression.
}
\description{
Estimate Eta Using LAD Regression
}
