% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{GetWeights}
\alias{GetWeights}
\title{Compute Kaplan-Meier Based Weights}
\usage{
GetWeights(data)
}
\arguments{
\item{data}{Data frame with 'time' and 'delta' columns.}
}
\value{
A numeric vector of weights.
}
\description{
Compute Kaplan-Meier Based Weights
}
