% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{RTransAFT}
\alias{RTransAFT}
\title{Refined RTransAFT Estimator via Adaptive LAD}
\usage{
RTransAFT(primdata, beta_a)
}
\arguments{
\item{primdata}{Primary dataset.}

\item{beta_a}{Auxiliary model coefficients.}
}
\value{
Estimated adjustment vector eta.
}
\description{
Refined RTransAFT Estimator via Adaptive LAD
}
