% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{runBtsp_RTransAFT_one}
\alias{runBtsp_RTransAFT_one}
\title{Bootstrap a Single Replicate for Refined Transfer AFT}
\usage{
runBtsp_RTransAFT_one(r)
}
\arguments{
\item{r}{Integer index for the bootstrap iteration.}
}
\value{
A list of estimated coefficient vectors: luad, seer, both, and transfer.
}
\description{
Bootstrap a Single Replicate for Refined Transfer AFT
}
